Fixed all noticeable platform bugs.
Platforms are now virtually impenetrable (can't enter from top, bottom, or sides). BitMan cannot walk through either side of the screen, and JOptionPane for difficulty has been fixed.
I found another issue. Whenever BitMan loses, he's supposed to go back to the top of the screen, right? He's not. He's staying next to the monster and losing all three lives practically at the same time.
Additionally, I agree with Kevin. Graphics are imperative for games, and the graphics for this game are mediocre at best. I suggest finding somebody who is good at art (with Photoshop or MS Paint, preferably PS), and ask them to draw some images that we could use. Not just for the background, but also for the characters. It would bring more life to the game, and also help with the coalescence of all the variables for BitMan (with images, you only need one x and y coordinate, not a million as we do now).